explicit cast to avoid a warning

+++ req/include/req_quantile_calculator_impl.hpp
@@ -49,7 +49,7 @@ void req_quantile_calculator<T, A>::convert_to_cummulative() {
 
 template<typename T, typename A>
 const T& req_quantile_calculator<T, A>::get_quantile(double rank) const {
-  uint64_t weight = rank * n_;
+  uint64_t weight = static_cast<uint64_t>(rank * n_);
   auto it = std::lower_bound(entries_.begin(), entries_.end(), Entry(nullptr, weight), compare_pairs_by_second());
   if (it == entries_.end()) return *(entries_[entries_.size() - 1].first);
   return *(it->first);
